+++ src/components/map-picker/MapExemple.vue
@@ -6,12 +6,13 @@
     <h3>Map Exemples :</h3>
 
     <h4> Controlled input value after selection : </h4>
-    <MapAutoComplete label="Enter a name of street" v-model="address" @onAddressSelected="onSelect" :formatSelectedValue="(option) => option.street"  />
+    <MapAutoComplete label="Enter a name of street" v-model="address" @onAddressSelected="onSelect"
+      :formatSelectedValue="(option) => option.street" />
     <h6> Selected : {{ address }}</h6>
     <br>
 
     <h4> Not Controlled input value after selection : </h4>
-    <MapAutoComplete label="Enter a name of street" v-model="address2"  />
+    <MapAutoComplete label="Enter a name of street" v-model="address2" />
     <h6> Selected : {{ address2 }}</h6>
 
     <br>
@@ -21,7 +22,8 @@
 
     <br>
     <h4> AutoCompleteMap with Radius set to 10 klm: </h4>
-    <MapAutoComplete v-model:radius="radius" label="Enter a name of street" v-model="address3" :defaultSelectedAddress="selectedAddress" />
+    <MapAutoComplete v-model:radius="radius" label="Enter a name of street" v-model="address3"
+      :defaultSelectedAddress="selectedAddress" />
     <h6> Selected : {{ address3 }} === Radius : {{ radius }}</h6>
 
 
@@ -62,7 +64,7 @@ export default {
   },
   methods: {
     onSelect(adr) {
-       
+
 
     }
   }